Seperating logic from biker controller to biker service and apply field validations

+++ src/Models/Biker.php
@@ -11,10 +11,10 @@ class Biker extends Model{
     protected string $updated_at;
 
     public static array $fieldRules = [
-        'id' => ['type' => 'int', 'min' => 0],
+        'id' => ['type' => 'integer', 'min' => 0],
         'name' => ['type' => 'string', 'min_length' => 1, 'max_length' => 255, 'required' => true],
-        'latitude' => ['type' => 'float', 'min' => -90, 'max' => 90, 'required' => true],
-        'longitude' => ['type' => 'float', 'min' => -180, 'max' => 180, 'required' => true],
+        'latitude' => ['type' => 'double', 'min' => -90, 'max' => 90, 'required' => true],
+        'longitude' => ['type' => 'double', 'min' => -180, 'max' => 180, 'required' => true],
         'updated_at' => ['type' => 'string', 'min_length' => 1, 'max_length' => 255, 'required' => true]
     ];
 }
